// src/core/commission/domain/interfaces/sort_commission.ts
import type { Order } from "@/src/core/app/domain/models/order";
export interface SortCommission {
  marketerId?: Order;
  name?: Order;
  priceType?: Order;
  energyType?: Order;
  rateType?: Order;
  rates?: Order;
  minPower?: Order;
  maxPower?: Order;
  minConsumption?: Order;
  maxConsumption?: Order;
  percentagetestCommission?: Order;
  testCommission?: Order;
  createdAt?: Order;
}